refactor: move color and padding utils in theme-utils
Change pseudoClasses utils to create css rules for a variable css property
and to accept a custom color
