--- src/components/utilities/functions.ts
@@ -1,47 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2021 Zextras <https://www.zextras.com>
- *
- * SPDX-License-Identifier: AGPL-3.0-only
- */
-
-import { css, FlattenSimpleInterpolation } from 'styled-components';
-import type { ThemeObj } from '../../theme/theme';
-
-function pseudoClasses(
-	theme: ThemeObj,
-	color: keyof ThemeObj['palette']
-): FlattenSimpleInterpolation {
-	return css`
-		transition: background 0.2s ease-out;
-		&:focus {
-			outline: none;
-			background: ${theme.palette[color].focus};
-		}
-		&:hover {
-			outline: none;
-			background: ${theme.palette[color].hover};
-		}
-		&:active {
-			outline: none;
-			background: ${theme.palette[color].active};
-		}
-		&:disabled {
-			outline: none;
-			background: ${theme.palette[color].disabled};
-		}
-	`;
-}
-
-function parsePadding(padding: string, theme: ThemeObj): string {
-	let paddingValue = padding;
-	const paddingSizes = Object.keys(theme.sizes.padding) as Array<
-		keyof ThemeObj['sizes']['padding']
-	>;
-	paddingSizes.forEach((size) => {
-		const regex = new RegExp(`(^|\\s)(${size})`, 'g');
-		paddingValue = paddingValue.replace(regex, `$1${theme.sizes.padding[size]}`);
-	});
-	return paddingValue;
-}
-
-export { pseudoClasses, parsePadding };
